correctly parse events with raw.event perserved
changed \= to (?<!\\)\=
the literal equals sign is too restrictive and greedy --  we need to make sure it is not preceded by a backslash to capture raw events contained within valid CEF.  Example:

ad.rawEvent=this is whole CEF event but escaped out like this destinationAddress\=1.1.1.1 sourceAddress\=2.2.2.2

also additional fields (ad.*) often contain brackets causing errors.  A bracket is interpreted as a nested array in logstash, which leads to messy events and parsing errors.  Simply swapping out the brackets with underscores works great.