Make cef.rb encoder compliant to spec.
The CEF specifications document stipulates that "When sending events, the CEF key name is the proper form to use" (this is the short name). The encoder was not compliant to this because it produced CEF messages with the "full name" instead of the "CEF key name". This modification introduces a reverse mapping for the fields that require it.