encoding: coerce input data to UTF-8 to comply with spec
The Common Event Format dicates that inputs MUST be `UTF-8`, but in Logstash
strings are typically passed to the Codec flagged as `BINARY` (regardless of
their contents).

Before processing the CEF input, use `LogStash::Util::Charset` to load the
input bytes as the CEF-specified UTF-8, ensuring that multi-byte character
operations are supported on the resulting strings downstream.

If doing so produces an invalid UTF-8 byte sequence, bail early to avoid
wasted work; this will cause the input bytes to be emitted verbatim as the
event's `message`.