encoding: coerce input data to UTF-8 to comply with spec
The Common Event Format dicates that inputs MUST be `UTF-8`, but in Logstash
strings are typically passed to the Codec flagged as `BINARY` (regardless of
their contents).

This patch attempts to force the encoding metadata of the string to UTF-8, but
if doing so would produce a string with invalid encoding, restores the
string's original encoding metadata and fails, causing an event to be emitted
with the verbatim data as its message.