Make cef.rb encoder compliant to spec. (#51)
* Make cef.rb encoder compliant to spec.

The CEF specifications document stipulates that "When sending events, the CEF key name is the proper form to use" (this is the short name). The encoder was not compliant to this because it produced CEF messages with the "full name" instead of the "CEF key name". This modification introduces a reverse mapping for the fields that require it.

* Update cef_spec.rb

Add test case for validation of CEF key names encoding into the short version.

* Applied changes suggested by colinsurprenant

Adjusted MAPPINGS and used MAPPINGS.invert for REVERSE_MAPPINGS.
Removed superfluous string interpolation.

* removed deviceMacAddress field (will be in separate pull request as suggested); added configuration option reverse_mapping and created a second test case to make sure we test both with and without reverse mappings.